More doc, new PWM period, timeout on receive packet for CRYPTO

+++ inc/motor.h
@@ -1,12 +1,66 @@
 #ifndef MOTOR_H_
 #define MOTOR_H_
 
-void Init_Motor();										// Init and configure both motors, start in forward mode
-void Motor_Start();										// Start forward 0% (you have to run this function before moving)
-void Motor_Stop();										// Stop both motors
-void Motor_changeDir(uint8_t forward);					// Change direction (backward 11 or forward 10)
-void Motor_Backward(uint8_t perPower);					// Change the speed of the wheels while going backward ( perPower : percent of the power we want)
-void Motor_Forward(uint8_t perPower);					// Change the speed of the wheels while going forward ( perPower : percent of the power we want)
-void Motor_setPower(float motorPower);					// Change the speed of the wheels (negative value is to go backward, positive to go forward)
+
+/* ---MOTOR CONFIGURATION-- */
+
+/**
+ * @brief	Init both motors
+ * @param	Nothing
+ * @return	Nothing
+ * @note	Nothing
+ */
+void Init_Motor();
+
+/**
+ * @brief	Start both motors
+ * @param	Nothing
+ * @return	Nothing
+ * @note	Nothing
+ */
+void Motor_Start();
+
+/**
+ * @brief	Stop both motors
+ * @param	Nothing
+ * @return	Nothing
+ * @note	Nothing
+ */
+void Motor_Stop();
+
+/**
+ * @brief	Change the direction of both motors
+ * @param	forward : FORWARD (10) for set the direction as forward and BACKWARD (11) for backward
+ * @return	Nothing
+ * @note	Nothing
+ */
+void Motor_changeDir(uint8_t forward);
+
+
+/* ---MOTOR MOVEMENT--- */
+
+/**
+ * @brief	Set both motors to move backward
+ * @param	perPower : percent (not really a percent because its on PERIOD_RESET_(R or L)MOTOR and not 100, PERIOD_RESET_(R or L)MOTOR = 5V) of power desired
+ * @return	Nothing
+ * @note	Nothing
+ */
+void Motor_Backward(uint16_t perPower);
+
+/**
+ * @brief	Set both motors to move forward
+ * @param	perPower : percent (not really a percent because its on PERIOD_RESET_(R or L)MOTOR and not 100, PERIOD_RESET_(R or L)MOTOR = 5V) of power desired
+ * @return	Nothing
+ * @note	Nothing
+ */
+void Motor_Forward(uint16_t perPower);
+
+/**
+ * @brief	Set the power of both motors
+ * @param	motorPower : percent (not really a percent because its on PERIOD_RESET_(R or L)MOTOR and not 100, PERIOD_RESET_(R or L)MOTOR = 5V) of power desired, can be negative
+ * @return	Nothing
+ * @note	negative value means go forward and a positive one go backward
+ */
+void Motor_setPower(float motorPower);
 
 #endif /* MOTOR_H_ */
